add background images and JS functionality
In the css file added in background images for the .drum class elements from the images folder. in the index.js folder created event listeners for both on click and keydown events. The "click" event listeners were focused on the .drum buttons, pulling the inner html and utilized in a switch case to pull the respective audio file to play in the browser. The keydown function used event.key to extract the corresponding value matching the button and then utilized the switch case to pull the right audio.